// src/features/complaints/models/FieldType.ts
export enum FieldType {
  TEXT = "TEXT",
  TEXTAREA = "TEXTAREA",

  NUMBER = "NUMBER",
  INTEGER = "INTEGER",
  DECIMAL = "DECIMAL",

  SINGLE_SELECT = "SINGLE_SELECT",
  MULTI_SELECT = "MULTI_SELECT",

  BOOLEAN = "BOOLEAN",

  DATE = "DATE",
  TIME = "TIME",
  DATETIME = "DATETIME",

  DURATION = "DURATION",
  SEVERITY = "SEVERITY",
  BODY_LOCATION = "BODY_LOCATION",

  IMAGE = "IMAGE",
  DRAWING = "DRAWING",
}